Hold teardown against every cancellation
`_await_teardown_shielded` shielded the first await and then re-awaited the
teardown bare. A shield protects only the await it wraps, so the retry was
unprotected: a second cancellation aimed at the caller landed on the
teardown itself, cancelling the SIGTERM/grace/SIGKILL escalation mid-flight.
The `suppress` around it then swallowed the evidence, and the docstring's
claim that teardown "completes regardless" was false — a SIGTERM-immune
stage could outlive the run that spawned it.

The wait now loops on a fresh `asyncio.shield` until the teardown future is
actually done, recording each cancellation and re-raising the first, so the
caller still sees exactly one. `cuprum.sh._execute_with_hooks` keeps its
single shield: it drains bookkeeping tasks rather than owning process
lifetimes, so a second cancellation there orphans nothing.

The new case in `test_pipeline_teardown_cancellation` cancels the waiter
twice inside an event-coordinated grace window and asserts the stage was
still reaped; it also asserts the escalation had not yet run when the second
cancellation landed, so it cannot pass by racing the clock. Restoring the
unshielded retry fails it.

The cancellation cases move out of `test_pipeline_timeouts` into that new
module: adding them there would have pushed the file past the 400-line cap,
and "what the caller sees when a deadline expires" and "what survives when
the caller cancels mid-teardown" are two concerns anyway.
