+++ cuprum/_process_lifecycle.py
@@ -84,33 +84,7 @@ async def _terminate_process_with_wait(
 
 
 async def _shielded_cleanup[T](cleanup: cabc.Awaitable[T]) -> T:
-    """Run ``cleanup`` to completion, holding off caller cancellation.
-
-    Cleanup runs in an owned task and every wait on it is shielded. A bare
-    ``await asyncio.shield(coro)`` is not enough on its own: the shield stops
-    the cancellation reaching the inner coroutine, but the *outer* task returns
-    immediately, so the caller propagates its ``CancelledError`` while cleanup
-    is still running — the tasks and processes it owns outlive the run.
-
-    Awaiting the task directly on the cancellation path is no better, because
-    cancelling a task propagates to whatever future it is blocked on: a second
-    ``cancel()`` landing on a bare ``await task`` cancels the cleanup itself.
-    The wait is therefore retried under a shield until the task reports done,
-    absorbing however many cancellations arrive, and only then re-raised.
-
-    Returns
-    -------
-    T
-        Whatever ``cleanup`` returned. Its failure propagates when the caller
-        was not cancelled — callers that must absorb cleanup failures say so
-        themselves, as :func:`_await_teardown_shielded` does.
-
-    Raises
-    ------
-    asyncio.CancelledError
-        Re-raised once the cleanup task has completed, however many
-        cancellations arrived while it ran.
-    """
+    """Complete cleanup before re-raising any caller cancellation."""
     task = asyncio.ensure_future(cleanup)
     try:
         return await asyncio.shield(task)
@@ -129,36 +103,7 @@ async def _shielded_cleanup[T](cleanup: cabc.Awaitable[T]) -> T:
 async def _await_teardown_shielded(
     teardowns: cabc.Iterable[cabc.Awaitable[object]],
 ) -> None:
-    """Run ``teardowns`` to completion, holding off caller cancellation.
-
-    Teardown runs in an owned, shielded task. Awaiting it directly would let a
-    caller cancelling the run interrupt it part-way: ``process.terminate()`` is
-    synchronous and would still have fired, but the cancellation lands on the
-    grace-period wait, so the ``SIGKILL`` escalation and the final reap never
-    run and a process ignoring ``SIGTERM`` survives the run that spawned it.
-
-    The cancellation is held until teardown finishes and then re-raised,
-    mirroring the ``asyncio.shield`` cleanup in
-    ``cuprum.sh._execute_with_hooks``.
-
-    :func:`_shielded_cleanup` owns the retry rules; were a further ``cancel()``
-    to reach the teardown task, ``gather`` would pass it on to each
-    ``_terminate_process``, skipping the ``SIGKILL`` escalation and the reap.
-
-    Termination always settles — it escalates to ``SIGKILL`` and awaits the
-    exit — so the wait is bounded by the grace period rather than by the
-    caller's patience.
-
-    Failures are absorbed via ``return_exceptions`` — every caller runs this
-    while another exception is already propagating, and teardown must not
-    replace it.
-
-    Raises
-    ------
-    asyncio.CancelledError
-        Re-raised after the shielded teardown completes, once the initial
-        cancellation that interrupted the shield has been held off.
-    """  # noqa: DOC502 - CancelledError propagates from _shielded_cleanup
+    """Complete teardown before re-raising any caller cancellation."""
     await _shielded_cleanup(asyncio.gather(*teardowns, return_exceptions=True))
 
 
@@ -166,24 +111,7 @@ async def _terminate_all_shielded(
     processes: cabc.Iterable[asyncio.subprocess.Process],
     cancel_grace: float,
 ) -> None:
-    """Terminate every process, holding off caller cancellation until done.
-
-    Termination owns its own task and is shielded. Awaiting it directly would
-    let a caller cancelling the run interrupt teardown: ``process.terminate()``
-    is synchronous and would still have fired, but the cancellation lands on
-    the grace-period wait, so the ``SIGKILL`` escalation and the final reap
-    never run and a process ignoring ``SIGTERM`` survives the run that spawned
-    it.
-
-    The cancellation is held until termination finishes and then re-raised,
-    mirroring the ``asyncio.shield`` cleanup in
-    ``cuprum.sh._execute_with_hooks``. Further cancellations arriving during
-    that wait are absorbed and the shielded wait retried, so no number of them
-    can strand a process — see :func:`_await_teardown_shielded`.
-
-    Failures are absorbed — every caller runs this while another exception is
-    already propagating, and teardown must not replace it.
-    """
+    """Terminate every process before re-raising caller cancellation."""
     await _await_teardown_shielded(
         _terminate_process(process, cancel_grace) for process in processes
     )
@@ -364,16 +292,7 @@ async def _terminate_timed_out_stages(
     pipeline cannot block waiting on a producer that is still running.
 
     Failures are absorbed — this runs while a timeout is already propagating,
-    and must not replace the ``TimeoutExpired`` the caller is waiting for.
-
-    Teardown owns its own task and is shielded from the caller. Awaiting the
-    terminations directly would let a caller cancelling ``Pipeline.run()``
-    interrupt them: ``process.terminate()`` is synchronous and would still have
-    fired, but the cancellation lands on the grace-period wait, so the
-    ``SIGKILL`` escalation and the final reap never run and a stage ignoring
-    ``SIGTERM`` survives. The cancellation is therefore held until termination
-    has finished, mirroring the ``asyncio.shield`` cleanup in
-    ``cuprum.sh._execute_with_hooks``.
+    and must not replace the ``TimeoutExpired`` the caller awaits.
     """
     await _terminate_all_shielded(processes, cancel_grace)
 
@@ -406,10 +325,6 @@ async def _terminate_pipeline_remaining_stages(
     terminating the remaining pipeline stages. This prevents pipelines from
     hanging on long-running producers/consumers when downstream work is no
     longer meaningful.
-
-    Teardown is shielded for the same reason as the timeout path: a caller
-    cancelling mid-grace must not strand a stage before its ``SIGKILL``
-    escalation runs.
 
     Returns the number of stages that were still running and so were
     terminated. The caller reports that count, which is how an operator tells
