+++ cuprum/unittests/test_pipeline_timeouts.py
@@ -23,19 +23,14 @@ import pytest
 
 from cuprum import ScopeConfig, TimeoutExpired, _pipeline_collect, scoped, sh
 from cuprum._pipeline_stream_results import _reconcile_pipe_tasks
-from cuprum._process_lifecycle import (
-    _shielded_cleanup,
-    _terminate_pipeline_remaining_stages,
-    _terminate_timed_out_stages,
-)
-from cuprum.sh import ExecutionContext, Pipeline, RunOutputOptions
+from cuprum._process_lifecycle import _shielded_cleanup
+from cuprum.sh import Pipeline, RunOutputOptions
 from tests.helpers.catalogue import python_catalogue
 from tests.helpers.execution import _RunKwargs
 from tests.helpers.timeouts import (
     child_argv,
     pending_tasks,
     started_pids,
-    stubborn_child_argv,
     wait_for_process_death,
 )
 
@@ -49,37 +44,6 @@ _PIPELINE_STAGES = 2
 
 # Long enough that the immune child reaches ``signal.signal`` before the
 # deadline reaps it, and short enough to keep the test quick.
-_STAGE_TIMEOUT = 1.0
-_CANCEL_GRACE = 1.0
-
-
-async def _poll_until(
-    predicate: cabc.Callable[[], bool],
-    message: str,
-    *,
-    seconds: float = 10.0,
-) -> None:
-    """Yield until ``predicate`` holds, failing rather than hanging.
-
-    Parameters
-    ----------
-    predicate:
-        Readiness check, polled until it returns true.
-    message:
-        Failure message reported if ``seconds`` elapses first.
-    seconds:
-        How long to wait before failing.
-    """
-    loop = asyncio.get_running_loop()
-    deadline = loop.time() + seconds
-    # ASYNC110: the readiness signal is a file written by another process,
-    # which no asyncio.Event can observe. Polling it is the coordination.
-    while not predicate():
-        if loop.time() > deadline:
-            pytest.fail(message)
-        await asyncio.sleep(0.01)
-
-
 # -- Public-boundary timeout contract -----------------------------------------
 #
 # The single-command counterparts live in ``test_safe_cmd_run``. A pipeline
@@ -188,256 +152,6 @@ def test_pipeline_non_positive_timeout_at_public_boundary(
         assert expired.stderr is None, (
             f"a non-capturing pipeline must leave stderr unset, got {detail}"
         )
-
-
-# -- Cancellation during post-timeout teardown --------------------------------
-
-
-class _SigtermImmuneProcess:
-    """Process double that ignores ``terminate()``; only ``kill()`` ends it."""
-
-    def __init__(self) -> None:
-        """Start unexited, recording which signals were delivered."""
-        self.returncode: int | None = None
-        self.pid = 9101
-        self.terminated = False
-        self.killed = False
-        self.terminate_signalled = asyncio.Event()
-        self.grace_started = asyncio.Event()
-        self._exited = asyncio.Event()
-
-    async def wait(self) -> int:
-        """Block until ``kill()`` records an exit code.
-
-        Entry is announced so a test can place a cancellation *inside* the
-        grace-period wait: ``terminate()`` has already returned by this point,
-        so the escalation to ``SIGKILL`` is the only thing still outstanding.
-
-        Returns
-        -------
-        int
-            The recorded exit code, or ``0`` when none was set.
-        """
-        self.grace_started.set()
-        await self._exited.wait()
-        return self.returncode if self.returncode is not None else 0
-
-    def terminate(self) -> None:
-        """Record the signal without exiting, as a SIGTERM-immune child would."""
-        self.terminated = True
-        self.terminate_signalled.set()
-
-    def kill(self) -> None:
-        """Record the escalation and exit."""
-        self.killed = True
-        self.returncode = -9
-        self._exited.set()
-
-
-def test_timeout_teardown_completes_despite_caller_cancellation() -> None:
-    """Cancelling mid-grace must not strand a stage before the SIGKILL escalation.
-
-    ``terminate()`` is synchronous and always lands, but the escalation waits
-    out ``cancel_grace`` first. If the caller's cancellation interrupted that
-    wait, a stage ignoring ``SIGTERM`` would survive the run that spawned it.
-    Teardown is therefore shielded and finishes before the cancellation
-    propagates.
-    """
-
-    async def run_case() -> None:
-        """Cancel the teardown inside its grace window and inspect the stage."""
-        process = _SigtermImmuneProcess()
-        task = asyncio.create_task(
-            _terminate_timed_out_stages(
-                [typ.cast("asyncio.subprocess.Process", process)], 0.2
-            )
-        )
-        await asyncio.sleep(0.02)
-        task.cancel()
-
-        with pytest.raises(asyncio.CancelledError):
-            await task
-
-        assert process.terminated, "teardown must still deliver the initial signal"
-        assert process.killed, (
-            "the SIGKILL escalation must complete before the cancellation "
-            "propagates, otherwise a SIGTERM-immune stage outlives the run"
-        )
-        assert process.returncode == -9, (
-            f"the stage must be reaped, got returncode={process.returncode!r}"
-        )
-
-    asyncio.run(run_case())
-
-
-def test_timeout_teardown_survives_repeated_cancellation() -> None:
-    """A second cancellation during the grace period must not strand a stage.
-
-    Cancelling a task propagates to whatever future it is awaiting, so a
-    further ``cancel()`` landing while teardown is being waited out would
-    cancel the teardown itself and skip the ``SIGKILL`` escalation. The wait is
-    therefore retried under a shield until it completes, however many
-    cancellations arrive.
-    """
-
-    async def run_case() -> None:
-        """Cancel twice inside the grace window and inspect the stage."""
-        process = _SigtermImmuneProcess()
-        task = asyncio.create_task(
-            _terminate_timed_out_stages(
-                [typ.cast("asyncio.subprocess.Process", process)], 0.2
-            )
-        )
-        # Block until teardown is actually waiting out the grace period, so
-        # both cancellations land inside it rather than before the initial
-        # signal was even delivered.
-        await process.grace_started.wait()
-
-        assert task.cancel(), "the first cancellation must reach a live teardown"
-        # One scheduler pass delivers that cancellation; the retry loop then
-        # re-parks on the shield within the same pass. The count is not what
-        # the test rests on — the assertion below is: `cancel()` returns False
-        # once a task has finished, so a second cancellation that is accepted
-        # proves the first did not end the teardown early.
-        await asyncio.sleep(0)
-        assert task.cancel(), (
-            "the teardown completed on the first cancellation, so the retry "
-            "path this test exists to cover was never exercised"
-        )
-
-        with pytest.raises(asyncio.CancelledError):
-            await task
-
-        assert process.killed, (
-            "the SIGKILL escalation must complete even when a second "
-            "cancellation lands during the grace-period wait"
-        )
-        assert process.returncode == -9, (
-            f"the stage must be reaped, got returncode={process.returncode!r}"
-        )
-
-    asyncio.run(run_case())
-
-
-def test_pipeline_run_cancelled_during_grace_still_reaps_stages(
-    tmp_path: Path,
-) -> None:
-    """Cancelling ``Pipeline.run()`` mid-teardown still leaves no stage running.
-
-    The end-to-end counterpart of the helper test above: a real first stage
-    ignores ``SIGTERM``, so only the escalation can reap it. Cancellation is
-    issued once the child has confirmed its handler is installed, so the test
-    does not race interpreter start-up.
-    """
-    catalogue, python_program = python_catalogue()
-    python = sh.make(python_program, catalogue=catalogue)
-    marker = tmp_path / "immune"
-    pipeline = python(*stubborn_child_argv(marker)) | python(
-        "-c", "import sys; sys.stdin.read()"
-    )
-    events: list[ExecEvent] = []
-
-    async def run_case() -> None:
-        """Time out, then cancel while the grace period is still running."""
-        ctx = ExecutionContext(cancel_grace=_CANCEL_GRACE)
-        loop = asyncio.get_running_loop()
-        started_at = loop.time()
-        task = asyncio.create_task(
-            pipeline.run(
-                timeout=_STAGE_TIMEOUT,
-                output=RunOutputOptions(capture=False),
-                context=ctx,
-            )
-        )
-        # The deadline has to outlast interpreter start-up. A ``SIGTERM``
-        # arriving before the child reaches ``signal.signal`` kills it outright,
-        # so it never writes the marker and never becomes the immune stage this
-        # test is about — and the wait below would then never finish.
-        await _poll_until(
-            marker.exists, "the immune first stage never signalled readiness"
-        )
-        # Teardown waits out ``cancel_grace`` after the deadline expires, and
-        # the cancellation belongs inside that window. There is no event to
-        # wait on: the ``timeout`` events are not emitted until teardown has
-        # already finished, so this is timed from the deadline itself.
-        delay = started_at + _STAGE_TIMEOUT + 0.2 - loop.time()
-        assert delay > 0, (
-            "the deadline elapsed before the child signalled readiness, so the "
-            "cancellation below would land after the grace window instead of "
-            f"inside it and quietly stop covering it (delay={delay!r})"
-        )
-        await asyncio.sleep(delay)
-        task.cancel()
-        with contextlib.suppress(asyncio.CancelledError):
-            await task
-
-    with (
-        scoped(ScopeConfig(allowlist=frozenset([python_program]))),
-        sh.observe(events.append),
-    ):
-        asyncio.run(run_case())
-
-    pids = started_pids(events)
-    assert pids, "the pipeline must report at least the first stage's pid"
-    for pid in pids:
-        wait_for_process_death(pid, context="cancellation during timeout teardown")
-
-
-# -- Cancellation during fail-fast teardown -----------------------------------
-
-
-def test_fail_fast_teardown_completes_despite_caller_cancellation() -> None:
-    """Fail-fast teardown is shielded like the timeout path.
-
-    ``_terminate_pipeline_remaining_stages`` tears down the surviving stages
-    after one exits non-zero. It runs inside the pipeline wait loop, so a
-    caller cancelling the run can land on its grace-period wait; without the
-    shield the ``SIGKILL`` escalation would be skipped and a ``SIGTERM``-immune
-    stage would outlive the pipeline.
-    """
-
-    async def run_case() -> None:
-        """Cancel fail-fast teardown mid-grace and inspect the surviving stage."""
-        failed = _SigtermImmuneProcess()
-        failed.returncode = 1
-        survivor = _SigtermImmuneProcess()
-        wait_tasks = [
-            asyncio.create_task(_settled(1)),
-            asyncio.create_task(survivor.wait()),
-        ]
-        await asyncio.sleep(0)
-
-        task = asyncio.create_task(
-            _terminate_pipeline_remaining_stages(
-                [
-                    typ.cast("asyncio.subprocess.Process", failed),
-                    typ.cast("asyncio.subprocess.Process", survivor),
-                ],
-                wait_tasks,
-                0,
-                cancel_grace=0.2,
-            )
-        )
-        await asyncio.sleep(0.02)
-        task.cancel()
-
-        with pytest.raises(asyncio.CancelledError):
-            await task
-
-        assert survivor.killed, (
-            "the SIGKILL escalation must complete before the cancellation "
-            "propagates, otherwise a SIGTERM-immune stage outlives the pipeline"
-        )
-        for wait_task in wait_tasks:
-            wait_task.cancel()
-
-    asyncio.run(run_case())
-
-
-async def _settled(code: int) -> int:
-    """Yield once, then return ``code``, standing in for an exited stage."""
-    await asyncio.sleep(0)
-    return code
 
 
 # -- Inter-stage pump ownership on the immediate path -------------------------
